Modified modal UI and behaviour
1. Added responsive breakpoints to change modal width
2. Added image to modal
3. Set scrollable-y behaviour for overflow
