+++ new_client/src/components/projects/IndividualProjectCard.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 import { makeStyles } from "@material-ui/core/styles";
 import Card from "@material-ui/core/Card";
 import CardActions from "@material-ui/core/CardActions";
@@ -7,7 +7,7 @@ import CardMedia from "@material-ui/core/CardMedia";
 import Button from "@material-ui/core/Button";
 import Typography from "@material-ui/core/Typography";
 import { overflowEllipsis } from "../Utilities/overflowEllipsis";
-import Modal from "@material-ui/core/Modal";
+import ProjectModal from "./ProjectModal";
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -24,46 +24,7 @@ const useStyles = makeStyles((theme) => ({
   font: {
     alignContent: "center",
   },
-  paper: {
-    position: "absolute",
-    width: "65%",
-    backgroundColor: "#FFF",
-    border: "2px solid #000",
-    boxShadow: theme.shadows[5],
-    padding: theme.spacing(2, 4, 3),
-    overflow: "scroll",
-  },
 }));
-
-const getModalStyle = () => {
-  const top = 50;
-  const left = 50;
-
-  return {
-    top: `${top}%`,
-    left: `${left}%`,
-    transform: `translate(-${top}%, -${left}%)`,
-  };
-};
-
-const ProjectModal = ({ open, modalHide, projectData }) => {
-  const classes = useStyles();
-  const [modalStyle] = React.useState(getModalStyle);
-
-  return (
-    <Modal open={open} onClose={modalHide}>
-      <div style={modalStyle} className={classes.paper}>
-        <Typography variant="h5" component="h2" align="center">
-          {projectData.title}
-        </Typography>
-        <hr />
-        <Typography variant="body1" component="p" align="center">
-          {projectData.desc}
-        </Typography>
-      </div>
-    </Modal>
-  );
-};
 
 export default function IndividualProjectCard({
   imgSrc,
@@ -119,7 +80,7 @@ export default function IndividualProjectCard({
             {displayDesc.isOverflow ? (
               <a
                 onClick={() => {
-                  handleShow({ title: title, desc: shortDesc });
+                  handleShow({ title: title, desc: shortDesc, image: imgSrc });
                 }}
               >
                 (read more)
